[FINNA-1977] Refactor marking of fees as paid to a common trait.
This change avoids multiple implementations for the same functionality.
